Move User class from UserAuthenticationTests --> Created a seperate file for the User class
