+++ Showcase-iOSTests/Login/Service/UserAuthenticationTests.swift
@@ -50,19 +50,6 @@ class UserAuthenticationTests: XCTestCase {
         verify(mockFirebaseAuthentication, times(1)).signIn(withEmail: anyString(), password: anyString(), completion: any())
     }
     
-    
-    class User: Equatable {
-        let email: String
-        
-        init(email: String) {
-            self.email = email
-        }
-        
-        static func == (lhs: UserAuthenticationTests.User, rhs: UserAuthenticationTests.User) -> Bool {
-            return lhs.email == rhs.email
-        }
-    }
-    
     enum AuthError: Error {
         case notAuthenticated
     }
